third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/iframe-popup-same-origin-to-unsafe-none.https-expected.txt

This is a testharness.js-based test.
PASS same-origin with SAME_ORIGIN iframe opening popup a SAME_ORIGIN with COOP: 
FAIL same-origin with SAME_SITE iframe opening popup a SAME_ORIGIN with COOP:  assert_equals: name expected "" but got "SAME_SITE_iframe_opening_SAME_ORIGIN_popup_with_coop_"
FAIL same-origin with CROSS_ORIGIN iframe opening popup a SAME_ORIGIN with COOP:  assert_equals: name expected "" but got "CROSS_ORIGIN_iframe_opening_SAME_ORIGIN_popup_with_coop_"
PASS same-origin with SAME_ORIGIN iframe opening popup a SAME_SITE with COOP: 
FAIL same-origin with SAME_SITE iframe opening popup a SAME_SITE with COOP:  assert_equals: name expected "" but got "SAME_SITE_iframe_opening_SAME_SITE_popup_with_coop_"
FAIL same-origin with CROSS_ORIGIN iframe opening popup a SAME_SITE with COOP:  assert_equals: name expected "" but got "CROSS_ORIGIN_iframe_opening_SAME_SITE_popup_with_coop_"
PASS same-origin with SAME_ORIGIN iframe opening popup a CROSS_ORIGIN with COOP: 
FAIL same-origin with SAME_SITE iframe opening popup a CROSS_ORIGIN with COOP:  assert_equals: name expected "" but got "SAME_SITE_iframe_opening_CROSS_ORIGIN_popup_with_coop_"
FAIL same-origin with CROSS_ORIGIN iframe opening popup a CROSS_ORIGIN with COOP:  assert_equals: name expected "" but got "CROSS_ORIGIN_iframe_opening_CROSS_ORIGIN_popup_with_coop_"
Harness: the test ran to completion.
